Move model controller.
